Refactor sql
