+++ core/src/main/scala/ai/lum/odinson/state/sql/IdProvider.scala
@@ -1,0 +1,11 @@
+package ai.lum.odinson.state.sql
+
+class IdProvider(protected var id: Int = 0) {
+
+  def next: Int = {
+    val result = id
+
+    id += 1
+    result
+  }
+}
