+++ core/src/main/scala/ai/lum/odinson/state/sql/ReadNode.scala
@@ -1,0 +1,3 @@
+package ai.lum.odinson.state.sql
+
+case class ReadNode(docIndex: Int, name: String, id: Int, parentId: Int, childCount: Int, childLabel: String, start: Int, end: Int)
